Refactor: "utils 구조변경"
